+++ NEWS.md
@@ -2,6 +2,11 @@
 
 ### Function changes
 - Add `imax` method to `run_ICA()`. This allows use of the `infomax` ICA algorithm from the `infomax` package, which is a reimplementation of the Infomax algorithm used in the `EEGLAB` Matlab toolbox.
+- `erp_scalp()` and `interactive_scalp()` should now appropriately use channel locations included in the data.
+
+### Internal changes / bug fixes
+
+- `erp_scalp()` and `interactive_scalp()` now use cleaner evaluation of the `colour` argument using `rlang` 
 
 # eegUtils 0.6.3